requirements.txt
streamlit
duckduckgo-search
trafilatura
requests
beautifulsoup4
pypdf
google-generativeai>=0.8.0